Add files via upload
I am writing to you because I officially have traced the URL code that claims to be the SB Courts portal that posted this.  According to the amendment of the privacy laws of the state of California Rule of the court which was revised September 1, 2019 in part reads: 2.503(c) A court that maintains the following records in electronic form must provide electronic access to them at the courthouse, to the extent it is feasible to do so, but may not provide public remote access to these records: (4) Records in mental health proceeding and section (5) Records in criminal proceeding; (10) records in an elder or dependent abuse prevention proceeding under Welfare and Institutions code section 15657.03.  In this case, First and foremost this certificate is NOT one that is provided to and authentic by the SB Courts electronic portal.  The authenticity of the SB Courts portal is not equivalent to one that the court system provides and this certificate should be immediately REVOKED based in part on the fact the URL when analyzed fails the compatibility analysis of the SB court system.  Rule 2.503(c) states in part the access to the criminal record is made available AT THE COURTHOUSE here, the malicious posting came from a location and posted it on the internet to an extent that requires no additional effort by the public to retrieve my record.  section (4) clearly shows a proceeding relative to the Clean and Sober court proceeding my case was referred to which falls under the "mental health" classification of state laws.  therefore section (c) has been violated.  Furthermore, section (5) refers to records in criminal proceedings, this is a record in criminal proceedings which is questionable as to whether or not the public gains a greater need for the information that I do as the negative effects it produces when seeking employment.  Although the court allows public access according to penal code section 1330, which generally prohibits a local criminal justice agency, including a court from distributing information that relates a person's criminal history.  It is reasonable to believe that although the public is entitled to access the individual court files, providing access in the form of the court's information system would permit the compilation and distribution of criminal histories, in violation of the statue. (Westbrook, supra, at pp. 163-165.) Based on these reasons alone I am requesting this posting be removed from the WWW, internet and anywhere else readily available other than the law provides.  This is detrimental to my ability to seek employment and I have researched enough to lead this back to Global Sign as a provider to this cert as well as the issued to: INCAPSULA is a internal name and after the SSL analysis report provided by Global Sign it is safe to conclude you are responsible for removing the content and access from the internet.  The benefits to the public outweigh the burden to my family and myself including but not limited to MY LIFE !!! Moreover, It is unlawful to have any posting of this nature based on the fact there was supposed to be a 5 day notice provided to both parties (thus the Santa Barbara Courthouse public defendants office and the public) before the court even makes a determination to provide remote access under this rule !!!! as Notice to the public may be accomplished by posting notice on the court's actual website.  Any person may file comments "with the court" for consideration but no hearing is required. In my case this was simply maliciously posted by an intruder as a matter of law.  Subdivision (c) states in part, " The court should not publish that information over the internet where it is public interest regarding a case that was widely disseminated through the media of non confidential court records, redacted where necessary to protect the privacy of the participants.  There should have been a judge that provided a determination on my case as to whether or not this can be made available on the internet and I should have received a copy of the order permitting remote access as well as the Criminal Justice Services Dept. located at 455 Golden Gate Avenue, San Francisco, CA  94102.  I NEVER received the said notice therefore I ask for this record to be removed from the internet IMMEDIATELY !!!! I will pursue further action if necessary.  At 7 years I can officially clear a background check, in so far as this record is concerned I will NEVER clear a background check and this alone violates my rights and the law.  !!!! Below is the URL the posting is derived from.  Please locate and remove immediatly !!! from ALL search engines other that the actual court portal which requires public registration.